+++ src/events.ts
@@ -13,8 +13,17 @@ const PREFIX = process.env.PREFIX || DEFAULT_PREFIX;
 cog.addEvent({
   name: Events.ClientReady,
   once: true,
-  async execute(...args: any) {
-    log(`Logged in as ${args[0].user.tag}!`, "READY");
+  async execute(client, modifiedClient: Client, { loadingLog, loadedCog }) {
+    log(`Logged in as ${client.user.tag}!`, loadingLog);
+    let loadedCogs = "";
+    modifiedClient.cogs.forEach((cog) => {
+      loadedCogs += `- ${cog.name} ✅\n`;
+    });
+
+    loadedCog.setContent("");
+    await new Promise((r) => setTimeout(r, 3000));
+    loadedCog.setContent(loadedCogs);
+    loadedCog.setLabel("Loaded Cogs " + `(${modifiedClient.cogs.size})`);
   },
 } as EventFunction);
 
@@ -22,7 +31,11 @@ cog.addEvent({
 cog.addEvent({
   name: Events.MessageCreate,
   once: false,
-  async execute(message: any, modifiedClient: Client) {
+  async execute(
+    message: any,
+    modifiedClient: Client,
+    { loadingLog, errorLog }
+  ) {
     if (message.author.bot) return;
 
     let messageArray = message.content.split(" ");
@@ -48,7 +61,7 @@ cog.addEvent({
     if (!command.execute) {
       log(
         `Command exists but no execute function found for "${commandName}"`,
-        "WARNING"
+        errorLog
       );
       return;
     }